maint: Break out token_backend from permissions
