+++ custom_components/ms365_todo/classes/permissions.py
@@ -1,22 +1,15 @@
 """Generic Permissions processes."""
 
 import logging
-import os
 from copy import deepcopy
-
-from O365 import FileSystemTokenBackend
 
 from ..const import (
     CONF_ENTITY_NAME,
-    MS365_STORAGE_TOKEN,
     TOKEN_ERROR_CORRUPT,
-    TOKEN_ERROR_MISSING,
     TOKEN_ERROR_PERMISSIONS,
     TOKEN_FILE_CORRUPTED,
     TOKEN_FILE_PERMISSIONS,
-    TOKEN_FILENAME,
 )
-from ..helpers.filemgmt import build_config_file_path
 from ..integration.const_integration import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
@@ -25,7 +18,7 @@ _LOGGER = logging.getLogger(__name__)
 class BasePermissions:
     """Class in support of building permission sets."""
 
-    def __init__(self, hass, config):
+    def __init__(self, hass, config, token_backend):
         """Initialise the class."""
         self._hass = hass
         self._config = config
@@ -33,13 +26,7 @@ class BasePermissions:
         self._requested_permissions = []
         self._permissions = []
         self.failed_permissions = []
-        self.token_filename = self.build_token_filename()
-        self.token_path = build_config_file_path(self._hass, MS365_STORAGE_TOKEN)
-        _LOGGER.debug("Setup token")
-        self.token_backend = FileSystemTokenBackend(
-            token_path=self.token_path,
-            token_filename=self.token_filename,
-        )
+        self.ha_token_backend = token_backend
 
     @property
     def requested_permissions(self):
@@ -67,7 +54,7 @@ class BasePermissions:
             _LOGGER.warning(
                 TOKEN_ERROR_PERMISSIONS,
                 ", ".join(self.failed_permissions),
-                self.token_filename,
+                self.ha_token_backend.token_filename,
                 self._config[CONF_ENTITY_NAME],
             )
             return TOKEN_FILE_PERMISSIONS
@@ -117,14 +104,10 @@ class BasePermissions:
 
         return False
 
-    def build_token_filename(self):
-        """Create the token file name."""
-        return TOKEN_FILENAME.format(DOMAIN, f"_{self._config.get(CONF_ENTITY_NAME)}")
-
     def _get_permissions(self):
         """Get the permissions from the token file."""
 
-        scopes = self.token_backend.get_token_scopes()
+        scopes = self.ha_token_backend.token_backend.get_token_scopes()
         if scopes is None:
             _LOGGER.warning(
                 TOKEN_ERROR_CORRUPT,
@@ -138,17 +121,3 @@ class BasePermissions:
             scopes[idx] = scope.removeprefix("https://graph.microsoft.com/")
 
         return False, scopes
-
-    def delete_token(self):
-        """Delete the token."""
-        full_token_path = os.path.join(self.token_path, self.token_filename)
-        if os.path.exists(full_token_path):
-            os.remove(full_token_path)
-
-    def check_token_exists(self):
-        """Check if token file exists.."""
-        full_token_path = os.path.join(self.token_path, self.token_filename)
-        if not os.path.exists(full_token_path) or not os.path.isfile(full_token_path):
-            _LOGGER.warning(TOKEN_ERROR_MISSING, full_token_path)
-            return False
-        return True
